manually add missing values for MD counties with apostrophes
There seems to be a bug in CSBS's API; all counties with an apostrophe
in the name are missing since Apr 14.  There are only 4 of them in the
country, but it happens that 3 of them are in Maryland, including one in
the populous DMV area.

Data taken from https://www.csbs.org/information-covid-19-coronavirus